laporan gabungan

+++ CentCoffee/database/migrations/2024_10_14_010003_create_tbahanbakus_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('tbahanbakus', function (Blueprint $table) {
             $table->char('kode_bahan_baku', 15)->primary();
             $table->string('nama_bahan_baku', 50);
-            $table->float('stok_bahan_baku');
+            $table->integer('stok_bahan_baku');
             $table->string('satuan_bahan_baku', 10);
             $table->date('tanggal_kadaluarsa_bahan_baku');
         });
